+++ packages/babel-jest/src/loadBabelConfig.ts
@@ -7,3 +7,14 @@
 
 // this is a separate file so it can be mocked in tests
 export {loadPartialConfig} from '@babel/core';
+
+import {
+  PartialConfig,
+  TransformOptions,
+  // @ts-expect-error: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/51741
+  loadPartialConfigAsync as asyncVersion,
+} from '@babel/core';
+
+export const loadPartialConfigAsync: (
+  options?: TransformOptions,
+) => Promise<Readonly<PartialConfig> | null> = asyncVersion;